refactor(documents): extract heading-analysis parsers from Document
Move the two nested HTMLParser subclasses (heading-level scanner + first-heading
inspector) out of the 637-line documents.py into core.heading_analysis as
module-level HeadingLevelScanner / HeadingInspector classes; drop the unused
Document._HEADING_TAGS class var. documents.py down to ~547 lines.

Pure relocation; all 849 tests green, ruff clean.
