+++ CHANGELOG.md
@@ -28,6 +28,7 @@ and the project follows [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 - **De-duplicated shared logic across backends and fragments.** Citation/footnote-identifier parsing (`_normalise_footnote_id`, `_split_citation_keys`, `_citation_keys_from_payload`, the DOI/citation regexes) and the `_find_image` / `_find_table` IR lookups, previously copied verbatim in both the LaTeX and Typst writers, now live in one shared `texsmith.writers._ir_queries` module. `BaseFragment.build_config` / `inject` gained concrete defaults (`config_cls.from_context` / `config.inject_into`), so the ten fragments that followed that convention no longer each restate it (`frame` / `geometry` keep their bespoke parsing). The `_coerce_string` / `_format_authors` helpers copied across the article/book/letter templates are now a single `TemplateContextHelpers` mixin in `texsmith.templates.common`.
 - Merged `texsmith.core.html_utils` (`strip_html_comments`) into `texsmith.adapters.html_utils` and deleted the dead `texsmith.core.templates.text` module (the live `_squash_blank_lines` lives on in `wrapper`); `_discover_template_variables` is now defined once in `context_usage` and imported by `wrapper`. Front-matter glossary parsing moved from `texsmith.core.conversion.glossary` to `texsmith.core.glossary` (it has no conversion-layer dependency and sits beside the other front-matter resolvers).
 - Extracted the babel / BCP-47 language tables and mappers out of the 1039-line `texsmith.core.templates.manifest` into a focused `texsmith.core.templates.languages` module (manifest down to ~930 lines); `runtime` now imports the alias table from there instead of reaching into `manifest`.
+- Moved the two heading-analysis `HTMLParser` subclasses out of `texsmith.core.documents` (637 → ~547 lines) into `texsmith.core.heading_analysis` (`HeadingLevelScanner`, `HeadingInspector`), and dropped the unused `Document._HEADING_TAGS` class variable.
 
 ### Removed
