+++ src/texsmith/core/documents.py
@@ -35,9 +35,8 @@ import contextlib
 import copy
 from dataclasses import dataclass, field
 from enum import Enum
-from html.parser import HTMLParser
 from pathlib import Path
-from typing import Any, ClassVar
+from typing import Any
 
 from ..adapters.markdown import (
     DEFAULT_MARKDOWN_EXTENSIONS,
@@ -53,6 +52,7 @@ from .conversion.inputs import (
     extract_front_matter_slots,
 )
 from .diagnostics import DiagnosticEmitter, NullEmitter
+from .heading_analysis import HeadingInspector, HeadingLevelScanner
 from .metadata import PressMetadataError, normalise_press_metadata
 from .templates.runtime import coerce_base_level
 
@@ -442,36 +442,12 @@ class Document:
                 self.slot_options.update(base_options)
         self._invalidate_prepared()
 
-    _HEADING_TAGS: ClassVar[set[str]] = {"h1", "h2", "h3", "h4", "h5", "h6"}
-
-    class _HeadingLevelScanner(HTMLParser):
-        __slots__ = ("minimum_level",)
-
-        def __init__(self) -> None:
-            super().__init__()
-            self.minimum_level: int | None = None
-
-        def handle_starttag(self, tag: str, attrs: list[tuple[str, str | None]]) -> None:
-            if not tag:
-                return
-            name = tag.lower()
-            if not name.startswith("h") or len(name) < 2 or not name[1].isdigit():
-                return
-            try:
-                level = int(name[1])
-            except ValueError:
-                return
-            if not 1 <= level <= 6:
-                return
-            if self.minimum_level is None or level < self.minimum_level:
-                self.minimum_level = level
-
     @classmethod
     def _resolve_heading_alignment(cls, html: str, strategy: TitleStrategy) -> int:
         if strategy is not TitleStrategy.KEEP:
             return 0
 
-        scanner = cls._HeadingLevelScanner()
+        scanner = HeadingLevelScanner()
         try:
             scanner.feed(html)
         finally:
@@ -482,62 +458,9 @@ class Document:
             return 0
         return minimum - 1
 
-    class _HeadingInspector(HTMLParser):
-        __slots__ = ("_depth", "_resolved", "first_level", "level_counts", "parts")
-
-        def __init__(self) -> None:
-            super().__init__()
-            self._depth = 0
-            self._resolved = False
-            self.first_level: int | None = None
-            self.level_counts: dict[int, int] = {}
-            self.parts: list[str] = []
-
-        def handle_starttag(self, tag: str, attrs: list[tuple[str, str | None]]) -> None:
-            name = tag.lower()
-            if not name.startswith("h") or len(name) < 2 or not name[1].isdigit():
-                if self._depth:
-                    self._depth += 1
-                return
-
-            try:
-                level = int(name[1])
-            except ValueError:
-                if self._depth:
-                    self._depth += 1
-                return
-
-            if not 1 <= level <= 6:
-                if self._depth:
-                    self._depth += 1
-                return
-
-            self.level_counts[level] = self.level_counts.get(level, 0) + 1
-            if self._resolved:
-                return
-
-            if self.first_level is None:
-                self.first_level = level
-                self._depth = 1
-                return
-
-            if self._depth:
-                self._depth += 1
-
-        def handle_endtag(self, tag: str) -> None:
-            if not self._depth:
-                return
-            self._depth -= 1
-            if self._depth == 0:
-                self._resolved = True
-
-        def handle_data(self, data: str) -> None:
-            if self._depth and not self._resolved:
-                self.parts.append(data)
-
     def _extract_promoted_title(self) -> tuple[str | None, bool]:
         """Return the promoted title and whether the heading should be dropped."""
-        inspector = self._HeadingInspector()
+        inspector = HeadingInspector()
         try:
             inspector.feed(self._html)
         finally:
@@ -555,7 +478,7 @@ class Document:
 
     def _first_heading_level(self) -> int | None:
         """Return the level of the first heading in the document, if any."""
-        inspector = self._HeadingInspector()
+        inspector = HeadingInspector()
         try:
             inspector.feed(self._html)
         finally:
